Disable incompatible self update mechanism for settings plugin
