+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/NewRefreshVersionsImpl.kt
@@ -3,6 +3,7 @@ package de.fayard.refreshVersions.core.internal
 import de.fayard.refreshVersions.core.*
 import de.fayard.refreshVersions.core.extensions.gradle.hasDynamicVersion
 import de.fayard.refreshVersions.core.extensions.gradle.isRootProject
+import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
@@ -71,27 +72,9 @@ internal suspend fun lookupVersionCandidates(
             }
         }
 
-        val selfUpdateAsync = async {
-
-            val moduleId = ModuleId(group = "de.fayard.refreshVersions", name = "refreshVersions")
-
-            val versionsFetchers = RefreshVersionsConfigHolder.settings.getDependencyVersionFetchers(
-                httpClient = httpClient,
-                dependencyFilter = { dependency ->
-                    dependency.group == moduleId.group && dependency.name == moduleId.name
-                }
-            ).toList()
-
-            val currentVersion = RefreshVersionsCorePlugin.currentVersion
-
-            DependencyWithVersionCandidates(
-                moduleId = moduleId,
-                currentVersion = currentVersion,
-                versionsCandidates = versionsFetchers.getVersionCandidates(
-                    currentVersion = Version(currentVersion),
-                    resultMode = resultMode
-                )
-            )
+        val selfUpdateAsync: Deferred<DependencyWithVersionCandidates>? = when {
+            RefreshVersionsConfigHolder.isSetupViaPlugin -> null
+            else -> async { LegacyBoostrapUpdatesFinder.getSelfUpdates(httpClient, resultMode) }
         }
 
         val gradleUpdatesAsync = async {
@@ -123,13 +106,13 @@ internal suspend fun lookupVersionCandidates(
             dependenciesWithHardcodedVersions = dependenciesWithHardcodedVersions,
             dependenciesWithDynamicVersions = dependenciesWithDynamicVersions,
             gradleUpdates = gradleUpdatesAsync.await(),
-            selfUpdates = selfUpdateAsync.await()
+            selfUpdatesForLegacyBootstrap = selfUpdateAsync?.await()
         )
         TODO("Check version candidates for the same key are the same, or warn the user with actionable details")
     }
 }
 
-private fun Settings.getDependencyVersionFetchers(
+internal fun Settings.getDependencyVersionFetchers(
     httpClient: OkHttpClient,
     dependencyFilter: (Dependency) -> Boolean
 ): Sequence<DependencyVersionsFetcher> = getDependencyVersionFetchers(
